Restructure a bit so it's easy to add multiple intake event handlers
